TypeScript support (fixed) (#12)
* Add TypeScript support

This patch adds support for compiling the body of the <script> tag which uses
either lang="ts" or lang="typescript" using TypeScript compiler instead of
Babel. Full backwards compatibility is maintained by making Babel a fallback
compiler if the lang attrib is not one of the 'ts', 'typescript', or 'babel'.

Locating and loading of the typescript configuration is deferred to tsconfig
package.

In order to unit-test the preprocessor, new fixture components are added which
use the class-based component format. vue-property-decorator package is added as
a development dependency to support this component.

Signed-off-by: Hajime Yamasaki Vukelic <hayavuk@gmail.com>

* Update yarn.lock to sync vue and vue-template-compiler versions

Signed-off-by: Hajime Yamasaki Vukelic <hayavuk@gmail.com>

* Clean up code

- Use reuqire() instead of tsconfig
- Use standard JS keys instead of strings in `transforms` object

Signed-off-by: Hajime Yamasaki Vukelic <hayavuk@gmail.com>