Support for Functional Components (#50)
* Added a check for functional components

If a component is functional and does not have a script tag, do not try to parse the non-existent script tag.

* Update index.js

* updating isafunctionalcomponent

doesn't seem to be doing its job

* can't reproduce on local

so have to console.log json and then copy to local for testing...

* syntax error

don't edit in vim editor its bad mmkay

* object not Object

gimme back the last 4 minutes of my life

* boom goes the dynamite

IT WAS AN OBJECT NOT A STRING

* sexy es6 arrow func

cuz the package I forked from seems to like that kinda style ;)

* sanity check

* need sleep

* Update index.js

* style: newlines

removed newlines

no breaking changes

* test(functional): add test for functional components

a test documenting Vue's inability to cope with functional
components. throws a warning every time Vue tries to mount
one. $el is not defined.

breaking changes: none, but in the future $el should be
defined, as it can't be used for testing if it's
undefined

* fix(test): behaves differently locally

test behaves differently on travis-ci than
on my local machine, so now there is a test
to document that

breaking changes none

* fix(test): typeof didn't work

tried to use typeof, decided to use
casting

no breaking changes

* fix(testing): removed context

context is needed for mounting functional components
but not as an option for a render function

no breaking changes

* fix(testing): iie -> function

was passing an expression not a function
to onClick handler

no breaking changes

* fix(testing): properly call props.onClick

1. need to pass props in context object to Vue.compile
2. need '@click="props.onClick('stuff')" VS
	'@click="() => props.onClick('stuff')"

No breaking changes

* test(local): can't reproduce locally

can't reproduce $el.querySelector being defined
locally, yet on travis-ci it's defined.
Without it being defined I can't debugger and
figure out what the mock function isn't being called.

Breaking changes: none; objeys all test criteria
except for testing that mockFn was called